requirements.txt
pytz
duckdb
authlib
slack_bolt
apscheduler
python-dotenv
beautifulsoup4